changed type from int to long for timestamp and duration values

+++ src/bookmark.h
@@ -3,10 +3,10 @@
 
 struct Bookmark
 {
-    Bookmark(const QString &name, int timestamp, int duration)
+    Bookmark(const QString &name, long timestamp, long duration)
         : name(name), timestamp(timestamp), duration(duration) {};
 
     QString name;
-    int timestamp;
-    int duration;
+    long timestamp;
+    long duration;
 };
