+++ src/bookmarksmodel.cpp
@@ -4,8 +4,8 @@
 #include <chrono>
 
 using namespace std::chrono_literals;
-static constexpr int BOOKMARK_MAX_TIMESTAMP = static_cast<int>(std::chrono::milliseconds(24h).count());
-static constexpr int BOOKMARK_MAX_DURATION = static_cast<int>(std::chrono::milliseconds(3h).count());
+static constexpr long BOOKMARK_MAX_TIMESTAMP = std::chrono::milliseconds(24h).count();
+static constexpr long BOOKMARK_MAX_DURATION = std::chrono::milliseconds(3h).count();
 
 
 BookmarksModel::BookmarksModel(QObject *parent)
@@ -20,8 +20,8 @@ void BookmarksModel::generate_bookmarks(int amount)
     m_bookmarks.reserve(amount);
     for (int i = 0; i < amount; i++) {
         auto name = QString("Bookmark %1").arg(i);
-        int timestamp = std::rand() % BOOKMARK_MAX_TIMESTAMP;
-        int duration = std::rand() % BOOKMARK_MAX_DURATION;
+        long timestamp = std::rand() % BOOKMARK_MAX_TIMESTAMP;
+        long duration = std::rand() % BOOKMARK_MAX_DURATION;
         m_bookmarks.emplace_back(name, timestamp, duration);
     }
     std::ranges::sort(
